# openapi/local-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CSGHub Lite Local API",
    "version": "0.1.0",
    "description": "Structured documentation for the local `/api/*` and `/v1/*` endpoints served by `csghub-lite`.\n\nMaintenance note: whenever you add, remove, or rename an `/api/*` or `/v1/*` route in `internal/server/routes.go`, update this file and run the OpenAPI sync test before merging."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current csghub-lite server"
    },
    {
      "url": "http://localhost:11435",
      "description": "Default local server"
    }
  ],
...
  "paths": {
...
    "/api/load": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Eagerly load a model",
        "description": "Loads a model into memory and optionally streams progress while conversions are running.",
        "operationId": "loadModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoadRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF",
                "stream": false,
                "num_ctx": 131072,
                "num_parallel": 1,
                "n_gpu_layers": 40,
                "cache_type_k": "q8_0",
                "cache_type_v": "q8_0",
                "dtype": "q8_0"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Load completed or streaming progress returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoadResponse"
                },
                "example": {
                  "status": "ready"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "data: {\"status\":\"loading Qwen/Qwen3-0.6B-GGUF\"}\n\ndata: {\"status\":\"converting\",\"step\":\"quantizing\",\"current\":1,\"total\":3}\n\ndata: {\"status\":\"ready\"}\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/stop": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Stop a running model",
        "operationId": "stopModel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StopRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model stopped.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                },
                "example": {
                  "status": "stopped"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/generate": {
      "post": {
        "tags": [
          "Inference"
        ],
        "summary": "Generate text",
        "description": "Generates text from a prompt. By default the endpoint streams tokens as Server-Sent Events unless `stream` is set to `false`.",
        "operationId": "generateText",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF",
                "prompt": "Write a haiku about local inference.",
                "stream": false,
                "options": {
                  "temperature": 0.7,
                  "max_tokens": 128
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated text or a streaming event sequence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "data: {\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"response\":\"Local\",\"done\":false,\"created_at\":\"2026-03-31T12:00:00Z\"}\n\ndata: {\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"response\":\" inference\",\"done\":false,\"created_at\":\"2026-03-31T12:00:00Z\"}\n\ndata: {\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"response\":\"\",\"done\":true,\"created_at\":\"2026-03-31T12:00:01Z\"}\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": [
          "Inference"
        ],
        "summary": "Generate chat completions",
        "description": "Supports local models and, when `source` is `cloud`, OpenCSG-backed chat models. Streaming responses are sent as SSE when `Accept: text/event-stream` or `X-CSGHUB-Stream: sse` is provided; otherwise streaming defaults to NDJSON.",
        "operationId": "chatCompletions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF",
                "messages": [
                  {
                    "role": "system",
                    "content": "You are a helpful assistant."
                  },
                  {
                    "role": "user",
                    "content": "What is 1 + 1?"
                  }
                ],
                "stream": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion output or streaming token events.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "data: {\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"message\":{\"role\":\"assistant\",\"content\":\"2\"},\"done\":false,\"created_at\":\"2026-03-31T12:00:00Z\"}\n\ndata: {\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"done\":true,\"created_at\":\"2026-03-31T12:00:00Z\"}\n\n"
              },
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/NDJSONStream"
                },
                "example": "{\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"message\":{\"role\":\"assistant\",\"content\":\"2\"},\"done\":false,\"created_at\":\"2026-03-31T12:00:00Z\"}\n{\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"message\":{\"role\":\"assistant\",\"content\":\"\"},\"done\":true,\"created_at\":\"2026-03-31T12:00:00Z\"}\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "Compatibility"
        ],
        "summary": "OpenAI-compatible chat completions",
        "description": "Compatibility layer for OpenAI Chat Completions. When `stream` is true the server streams SSE chunks and ends with `data: [DONE]`. If `stream` is omitted, deployments can opt into streaming by starting the server with `--openai-stream-default` or setting `CSGHUB_LITE_OPENAI_STREAM_DEFAULT=true`.",
        "operationId": "openAIChatCompletions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenAIChatRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF",
                "messages": [
                  {
                    "role": "user",
                    "content": "Write a short haiku."
                  }
                ],
                "stream": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion output or streamed chunk sequence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIChatResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "data: {\"id\":\"chatcmpl-1\",\"object\":\"chat.completion.chunk\",\"created\":1710000000,\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"choices\":[{\"index\":0,\"delta\":{\"role\":\"assistant\",\"content\":\"Hello\"}}]}\n\ndata: {\"id\":\"chatcmpl-1\",\"object\":\"chat.completion.chunk\",\"created\":1710000000,\"model\":\"Qwen/Qwen3-0.6B-GGUF\",\"choices\":[{\"index\":0,\"delta\":{\"role\":\"assistant\",\"content\":\"\"},\"finish_reason\":\"stop\"}]}\n\ndata: [DONE]\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/OpenAIBadRequest"
          },
          "401": {
            "$ref": "#/components/responses/OpenAIUnauthorized"
          },
          "429": {
            "$ref": "#/components/responses/OpenAITooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/OpenAIInternalError"
          }
        }
      }
    },
...
    "/v1/responses": {
      "get": {
        "tags": [
          "Compatibility"
        ],
        "summary": "Responses API probe guard",
        "description": "Returns a clear compatibility error for websocket probes and other unsupported GET requests to the Responses API endpoint.",
        "operationId": "getResponsesUnsupported",
        "responses": {
          "426": {
            "description": "WebSocket transport is not supported on this endpoint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenAIErrorResponse"
                },
                "example": {
                  "error": {
                    "message": "websocket transport is not supported on this endpoint",
                    "type": "invalid_request_error"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Compatibility"
        ],
        "summary": "OpenAI-compatible Responses API",
        "description": "Minimal Responses API compatibility for Codex and OpenAI SDK clients. When `stream` is true the route emits Responses API-style SSE events and ends with `event: done` plus `[DONE]`.",
        "operationId": "openAIResponses",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenAIResponsesRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF",
                "input": "Summarize the latest local logs.",
                "stream": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Responses API output or streamed event sequence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsesResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "event: response.created\ndata: {\"type\":\"response.created\",\"response\":{\"id\":\"resp_1\",\"object\":\"response\",\"status\":\"in_progress\"}}\n\nevent: response.completed\ndata: {\"type\":\"response.completed\",\"response\":{\"id\":\"resp_1\",\"object\":\"response\",\"status\":\"completed\"}}\n\nevent: done\ndata: [DONE]\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/OpenAIBadRequest"
          },
          "401": {
            "$ref": "#/components/responses/OpenAIUnauthorized"
          },
          "429": {
            "$ref": "#/components/responses/OpenAITooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/OpenAIInternalError"
          }
        }
      }
    },
    "/v1/messages": {
      "post": {
        "tags": [
          "Compatibility"
        ],
        "summary": "Anthropic-compatible messages API",
        "description": "Anthropic Messages compatibility endpoint. When `stream` is true the route emits Anthropic event-stream events such as `message_start`, `content_block_delta`, and `message_stop`.",
        "operationId": "anthropicMessages",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnthropicMessageRequest"
              },
              "example": {
                "model": "Qwen/Qwen3-0.6B-GGUF",
                "messages": [
                  {
                    "role": "user",
                    "content": "Explain what local inference means."
                  }
                ],
                "max_tokens": 256,
                "stream": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Anthropic-compatible message response or SSE event sequence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnthropicMessageResponse"
                }
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventStream"
                },
                "example": "event: message_start\ndata: {\"type\":\"message_start\",\"message\":{\"id\":\"msg_1\",\"type\":\"message\"}}\n\nevent: content_block_delta\ndata: {\"type\":\"content_block_delta\",\"index\":0,\"delta\":{\"type\":\"text_delta\",\"text\":\"Hello\"}}\n\nevent: message_stop\ndata: {\"type\":\"message_stop\"}\n\n"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AnthropicBadRequest"
          },
          "401": {
            "$ref": "#/components/responses/AnthropicUnauthorized"
          },
          "429": {
            "$ref": "#/components/responses/AnthropicTooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/AnthropicInternalError"
          }
        }
      }
    },
    "/v1/messages/count_tokens": {
      "post": {
        "tags": [
          "Compatibility"
        ],
        "summary": "Anthropic-compatible token counting",
        "operationId": "anthropicCountTokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnthropicCountTokensRequest"
              },
              "example": {
                "messages": [
                  {
                    "role": "user",
                    "content": "Count the tokens in this sentence."
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estimated input token count.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnthropicCountTokensResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/AnthropicBadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "BadRequest": {
        "description": "The request body, query parameters, or path parameters were invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required or the provided token is invalid.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "NotFound": {
        "description": "The requested model, dataset, app, or shell session was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "An internal server error occurred.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "OpenAIBadRequest": {
        "description": "The request was invalid for the OpenAI-compatible endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OpenAIErrorResponse"
            }
          }
        }
      },
      "OpenAIUnauthorized": {
        "description": "Authentication is required or invalid for the OpenAI-compatible endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OpenAIErrorResponse"
            }
          }
        }
      },
      "OpenAITooManyRequests": {
        "description": "The OpenAI-compatible endpoint rejected the request due to rate limiting.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OpenAIErrorResponse"
            }
          }
        }
      },
      "OpenAIInternalError": {
        "description": "The OpenAI-compatible endpoint encountered an internal error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/OpenAIErrorResponse"
            }
          }
        }
      },
      "AnthropicBadRequest": {
        "description": "The request was invalid for the Anthropic-compatible endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnthropicErrorResponse"
            }
          }
        }
      },
      "AnthropicUnauthorized": {
        "description": "Authentication is required or invalid for the Anthropic-compatible endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnthropicErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "EventStream": {
        "type": "string",
        "description": "Server-Sent Event stream payload."
      },
...
      "NDJSONStream": {
        "type": "string",
        "description": "Newline-delimited JSON stream payload."
      },
...
      "StatusResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "GenerateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "model",
          "prompt"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model ID to run. Local models accept the short public ID returned by model list APIs; legacy `namespace/name` IDs are also accepted."
          },
          "prompt": {
            "type": "string"
          },
          "stream": {
            "type": "boolean",
            "default": true
          },
          "options": {
            "$ref": "#/components/schemas/ModelOptions"
          },
          "web_search": {
            "$ref": "#/components/schemas/WebSearchOptions"
          }
        }
      },
      "GenerateResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "model",
          "response",
          "done",
          "created_at"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model ID returned by the request."
          },
          "response": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ChatRequest": {
        "type": "object",
        "required": [
          "model",
          "messages"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model ID to run. Local models accept the short public ID returned by model list APIs; legacy `namespace/name` IDs are also accepted."
          },
          "source": {
            "type": "string",
            "description": "Optional source hint. Use `cloud` to force an OpenCSG-backed chat model."
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tool"
            }
          },
          "stream": {
            "type": "boolean",
            "default": true
          },
          "options": {
            "$ref": "#/components/schemas/ModelOptions"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "model",
          "done",
          "created_at"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model ID returned by the request."
          },
          "message": {
            "$ref": "#/components/schemas/Message"
          },
          "done": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "StopRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "model"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model ID to stop. Local models accept the short public ID returned by model list APIs; legacy `namespace/name` IDs are also accepted."
          }
        }
      },
      "LoadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "model"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "Model ID to load. Local models accept the short public ID returned by model list APIs; legacy `namespace/name` IDs are also accepted."
          },
          "stream": {
            "type": "boolean",
            "default": false
          },
          "keep_alive": {
            "type": "string"
          },
          "num_ctx": {
            "type": "integer",
            "minimum": 1
          },
          "num_parallel": {
            "type": "integer",
            "minimum": 1
          },
          "n_gpu_layers": {
            "type": "integer",
            "minimum": 0
          },
          "cache_type_k": {
            "type": "string"
          },
          "cache_type_v": {
            "type": "string"
          },
          "dtype": {
            "type": "string"
          }
        }
      },
      "LoadResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          },
          "step": {
            "type": "string"
          },
          "current": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "OpenAIErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "errorCode"
        ],
        "properties": {
          "errorCode": {
            "type": "integer",
            "format": "int32",
            "example": 500
          },
          "error": {
            "$ref": "#/components/schemas/OpenAIError"
          }
        }
      },
...
      "OpenAIChatRequest": {
        "type": "object",
        "required": [
          "model",
          "messages"
        ],
        "properties": {
          "model": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "description": "Model source such as local, cloud, or provider:<id>."
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tool"
            }
          },
          "tool_choice": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ]
          },
          "parallel_tool_calls": {
            "type": "boolean"
          },
          "chat_template_kwargs": {
            "type": "object",
            "additionalProperties": true,
            "description": "llama.cpp chat template variables, such as {\"enable_thinking\": false} for Qwen thinking models."
          },
          "stream": {
            "type": "boolean",
            "default": false,
            "description": "Whether to stream SSE chunks. When omitted, the server startup configuration controls the default; the standard default is false."
          },
          "temperature": {
            "type": "number"
          },
          "top_p": {
            "type": "number"
          },
          "max_tokens": {
            "type": "integer"
          },
          "num_ctx": {
            "type": "integer"
          },
          "num_parallel": {
            "type": "integer"
          },
          "n_gpu_layers": {
            "type": "integer"
          },
          "cache_type_k": {
            "type": "string"
          },
          "cache_type_v": {
            "type": "string"
          },
          "dtype": {
            "type": "string"
          },
          "seed": {
            "type": "integer"
          },
          "stop": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "OpenAIChatResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "object",
          "created",
          "model",
          "choices",
          "usage"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "created": {
            "type": "integer"
          },
          "model": {
            "type": "string"
          },
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpenAIChoice"
            }
          },
          "usage": {
            "$ref": "#/components/schemas/OpenAIUsage"
          }
        }
      },
...
      "OpenAIResponsesRequest": {
        "type": "object",
        "required": [
          "model"
        ],
        "properties": {
          "model": {
            "type": "string"
          },
          "input": {},
          "instructions": {
            "type": "string"
          },
          "stream": {
            "type": "boolean",
            "default": false
          },
          "max_output_tokens": {
            "type": "integer"
          },
          "temperature": {
            "type": "number"
          },
          "top_p": {
            "type": "number"
          },
          "tools": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "tool_choice": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ]
          },
          "parallel_tool_calls": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}